jingjai_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
psycopg[binary]
passlib
bcrypt
python-jose
python-multipart
email-validator
pydantic
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
pyhton-dotenv
